backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
pydantic
python-multipart
python-jose[cryptography]